requirements.txt
langchain
langchain-community
langchain-openai
faiss-cpu
pypdf
unstructured
python-docx
sentence-transformers
